+++ Release_Notes.md
@@ -8,6 +8,7 @@
 - Converted the Dashboard "skin" to standalone
 - Lazy loading has been implemented properly and moved back to the gw-front-end project
     Added a readme.md to the gw-front-end/src/app/routes folder
+- Forgot password's cancel now closes the modal and opens a new one for login
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure